Silly update, forgot to compile locally before compiling remote.
Now that we are using akka 1.2 the library is throwing warnings about the deprecated behavior. These warnings are telling me that the return type of our previous message sending was an "illusion". It appears a cast to expected type used to happen inside the future processing. This cast was actually not ok because it wasn't proven that for all possible situations this cast would be correct. As a result when the unexpected happens the library throws class cast errors.

The fix in 1.2 is to stop making this inner cast and instead require the user to explicitly cast the future result with .as[T]. This way the location of type errors is made clear.

I'm glad I decided to make this switch now, because this is exactly the behavior I'm trying to troubleshoot.